+++ python/acl_anthology/anthology.py
@@ -411,8 +411,13 @@ class Anthology:
     ) -> list[Person]:  # pragma: no cover
         ...
 
-    def resolve(self, name_spec: NameSpecificationOrIter) -> PersonOrList:
+    def resolve(
+        self, name_spec: NameSpecificationOrIter
+    ) -> PersonOrList:  # pragma: no cover
         """Resolve a name specification (e.g. as attached to papers) to a natural person.
+
+        Warning:
+            Deprecated in favor of [`NameSpecification.resolve()`][acl_anthology.people.name.NameSpecification.resolve]; alternatively, [`PersonIndex.get_by_namespec()`][acl_anthology.people.index.PersonIndex.get_by_namespec] if you want to see what a hypothetical NameSpecification would resolve to that is not yet attached to a paper.
 
         Parameters:
             name_spec: A name specification, or an iterator over name specifications.
@@ -425,6 +430,12 @@ class Anthology:
             >>> anthology.resolve(paper.authors)
             [Person(id='lauri-karttunen', ...), Person(id='ronald-kaplan', ...), Person(id='annie-zaenen', ...)]
         """
+        warnings.warn(
+            DeprecationWarning(
+                "Anthology.resolve() is deprecated in favor of NameSpecification.resolve()"
+            )
+        )
+
         if isinstance(name_spec, NameSpecification):
             return self.people.get_by_namespec(name_spec)
         return [self.people.get_by_namespec(ns) for ns in name_spec]