+++ python/acl_anthology/people/name.py
@@ -14,13 +14,20 @@
 
 from __future__ import annotations
 
-from attrs import define, field, validators as v
+from attrs import define, field, setters, validators as v
 from functools import cache, cached_property
 from lxml import etree
 from lxml.builder import E
 import re
 from slugify import slugify
-from typing import Any, Optional, cast, TypeAlias
+from typing import Any, Optional, cast, TypeAlias, TYPE_CHECKING
+import sys
+
+if sys.version_info >= (3, 11):
+    from typing import Self
+else:
+    from typing_extensions import Self
+
 import yaml
 
 try:
@@ -28,7 +35,43 @@ try:
 except ImportError:  # pragma: no cover
     from yaml import Dumper  # type: ignore
 
+from ..exceptions import AnthologyException
+from ..utils.attrs import track_namespec_modifications
 from ..utils.latex import latex_encode
+
+if TYPE_CHECKING:
+    from ..anthology import Anthology
+    from ..collections import Volume, Paper, Talk
+    from ..people import Person
+
+
+SLUGIFY_REPLACEMENTS = (
+    ["ʼ", "-"],
+    ["’", "-"],
+)
+"""Custom replacement rules for name slugs."""
+
+
+LAST_NAME_LOWERCASE_PREFIXES = {
+    "al",
+    "bin",
+    "bint",
+    "da",
+    "de",
+    "del",
+    "di",
+    "dos",
+    "du",
+    "la",
+    "le",
+    "van",
+    "von",
+}
+"""Strings that tend to be lowercased when prefixing a last name; used for [`NameSpecification.normalize()`][acl_anthology.people.name.NameSpecification.normalize]."""
+
+
+LAST_NAME_CAPITALIZATION_RULES = ((r"^Mc([a-z])", lambda p: "Mc" + p.group(1).upper()),)
+"""Regex rules for heuristically normalizing last names; used for [`NameSpecification.normalize()`][acl_anthology.people.name.NameSpecification.normalize]."""
 
 
 @define(frozen=True)
@@ -126,7 +169,7 @@ class Name:
         Returns:
             A [slugified string](https://github.com/un33k/python-slugify#how-to-use) of the full name.
         """
-        return slugify(self.as_first_last())
+        return slugify(self.as_first_last(), replacements=SLUGIFY_REPLACEMENTS)
 
     @classmethod
     def from_dict(cls, name: dict[str, str]) -> Name:
@@ -246,13 +289,16 @@ def _Name_from(value: Any) -> Name:
     return Name.from_(value)
 
 
-@define
+@define(
+    on_setattr=[setters.convert, setters.validate, track_namespec_modifications],
+)
 class NameSpecification:
     """A name specification on a paper etc., containing additional data fields for information or disambiguation besides just the name.
 
     Attributes:
         name: The person's name.
         id: Unique ID for the person that this name refers to.
+        parent: The Anthology item that this name specification belongs to.
         orcid: An ORCID that was supplied together with this name.
         affiliation: Professional affiliation.
         variants: Variant spellings of this name in different scripts.
@@ -266,6 +312,7 @@ class NameSpecification:
 
     name: Name = field(converter=_Name_from)
     id: Optional[str] = field(default=None, validator=v.optional(v.instance_of(str)))
+    parent: Optional[Paper | Volume | Talk] = field(default=None, repr=False, eq=False)
     orcid: Optional[str] = field(default=None, validator=v.optional(v.instance_of(str)))
     affiliation: Optional[str] = field(
         default=None, validator=v.optional(v.instance_of(str))
@@ -291,12 +338,73 @@ class NameSpecification:
         """The last name component."""
         return self.name.last
 
+    @property
+    def root(self) -> Anthology:
+        """The Anthology instance to which this object belongs."""
+        if self.parent is None:  # pragma: no cover
+            raise AnthologyException(
+                "NameSpecification is not attached to an Anthology item."
+            )
+        return self.parent.root
+
     @cached_property
     def citeproc_dict(self) -> dict[str, str]:
         """A citation object corresponding to this name for use with CiteProcJSON."""
         if not self.name.first:
             return {"family": self.name.last}
         return {"family": self.name.last, "given": self.name.first}
+
+    def normalize(self) -> Self:
+        """Try to heuristically normalize the name.
+
+        Concretely, this will:
+        - Normalize the casing of the name *iff* it is currently all-lowercased or all-uppercased.
+        - Check against known verified names to see if a normalized variant already exists, and choose that one.
+        """
+        first, last = self.name.first, self.name.last
+
+        # Casing heuristics
+        firstlast = self.name.as_first_last()
+        if firstlast.islower() or firstlast.isupper():
+            if first is not None:
+                first = first.title()
+            last = last.title()
+            last_parts = last.split(" ")
+            # Prefixes
+            if (
+                len(last_parts) > 1
+                and last_parts[0].lower() in LAST_NAME_LOWERCASE_PREFIXES
+            ):
+                last = last[0].lower() + last[1:]
+            # Other normalization rules
+            for pattern, substitute in LAST_NAME_CAPITALIZATION_RULES:
+                last = re.sub(pattern, substitute, last)
+
+        # Checks against verified persons
+        name = Name(first, last)
+        for pid in self.root.people.slugs_to_verified_ids.get(name.slugify(), []):
+            canonical = self.root.people[pid].canonical_name
+            if canonical.slugify() == name.slugify():
+                name = canonical
+                break
+
+        self.name = name
+        return self
+
+    def resolve(self) -> Person:
+        """Resolve this name specification to a natural person.
+
+        Returns:
+            The Person object that this name specification resolves to.
+
+        Raises:
+            AnthologyException: If this name specification is not attached to an Anthology item (i.e., `parent` is not set).
+        """
+        if self.parent is None:
+            raise AnthologyException(
+                "Cannot resolve NameSpecification that is not attached to a paper."
+            )
+        return self.parent.root.people.get_by_namespec(self)
 
     @classmethod
     def from_xml(cls, person: etree._Element) -> NameSpecification: