+++ python/CHANGELOG.md
@@ -5,11 +5,18 @@
 ### Added
 
 - Files (such as paper PDFs) can now be downloaded from their remote URLs via `.download()`.
+- Added `.get_namespec_for(Person)` on papers and volumes, to more easily find the NameSpecification referring to a given Person.
+- Added `Person.namespecs()` to iterate over all NameSpecifications referring to this Person.
+- Added `NameSpecification.normalize()` to heuristically fix casing and match spelling details to known canonical names. This is mainly intended for ingestion and called automatically when using `create_paper()`/`create_volume()`.
 
 ### Changed
 
 - Switched tooling from Poetry to uv, and from Black to Ruff formatter.
 - Renamed `PersonIndex.resolve_namespec` to `_resolve_namespec` to discourage external use.
+- Improvements to Names and NameSpecifications:
+  - `Name.slugify()` now treats typographic apostrophes (U+02BC and U+2019) the same as regular ones.
+  - NameSpecifications now track if they have been modified, and will trigger their parent collection being saved on `Anthology.save_all()`.
+  - NameSpecifications can now be resolved via `NameSpecification.resolve()`. Therefore, `Anthology.resolve()` has been deprecated.
 
 ## [1.0.0] — 2026-01-24